tcintercom/requirements.txt
aiofiles==23.2.1
aioredis==2.0.1
arq==0.26.0
beautifulsoup4==4.12.3
fastapi==0.111.0
fastapi_utilities==0.2.0
httpx==0.27.0
pydantic==2.7.2
pydantic-settings==2.2.1
PyJWT==2.8.0
requests==2.32.3
sentry_sdk==2.3.1
starlette==0.37.2
uvicorn==0.30.1
